Crate tiktoken_rs

Rust library for tokenizing text for GPT using tiktoken

Counting token length

use tiktoken_rs::tiktoken::p50k_base;

let bpe = p50k_base().unwrap();
let tokens = bpe.encode_with_special_tokens(
  "This is a sentence   with spaces"
);
println!("Token count: {}", tokens.len());

Examples

For full working examples for all supported features see examples directory in the repository.

Modules

tiktoken

Adaptation of the tiktoken crate for use in Rust projects