Module tiktoken_rs::tiktoken

Adaptation of the tiktoken crate for use in Rust projects

Structs

CoreBPE

Functions

byte_pair_encode

byte_pair_split

cl100k_base

Use for ChatGPT models, text-embedding-ada-002 Initializes and returns a new instance of the cl100k_base tokenizer.

cl100k_base_singleton

Returns a singleton instance of the cl100k_base tokenizer. Use for ChatGPT models, text-embedding-ada-002

p50k_base

Use for Code models, text-davinci-002, text-davinci-003 Initializes and returns a new instance of the p50k_base tokenizer.

p50k_base_singleton

Returns a singleton instance of the p50k_base tokenizer. Use for Code models, text-davinci-002, text-davinci-003

r50k_base

Use for GPT-3 models like davinci Initializes and returns a new instance of the r50k_base tokenizer (also known as gpt2)

r50k_base_singleton

Returns a singleton instance of the r50k_base tokenizer. (also known as gpt2) Use for GPT-3 models like davinci